refactor: PieceImpl 패키지 재정의
