+++ src/main/java/chess/domain/PieceImpl/AbstractPiece.java
@@ -1,0 +1,7 @@
+package chess.domain.PieceImpl;
+
+import chess.domain.Piece;
+
+public abstract class AbstractPiece implements Piece {
+
+}
